# doc/swagger.yaml
openapi: 3.0.0
paths:
...
  /user:
    post:
      operationId: UserController_create
      summary: Create user
      description: Creates a new user
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserDto'
      responses:
        '201':
          description: The user has been created
          content:
            application/json:
              example:
                id: b1b73593-2445-421a-af42-359114d6c536
                login: string
                role: admin
                createdAt: 1775287579998
                updatedAt: 1775287579998
        '400':
          description: Bad request. Body does not contain required fields
        '403':
          description: Insufficient permissions
      tags:
        - Users
    get:
      operationId: UserController_findAll
      summary: Get all users
      description: Gets all users. Supports pagination and sorting
      parameters:
        - name: page
          required: false
          in: query
          description: Page number (starting from 1)
          schema:
            example: 1
            type: number
        - name: limit
          required: false
          in: query
          description: Number of items per page
          schema:
            example: 10
            type: number
        - name: sortBy
          required: false
          in: query
          description: The field to sort by
          schema:
            example: login
            enum:
              - id
              - login
              - role
              - createdAt
              - updatedAt
            type: string
        - name: order
          required: false
          in: query
          description: Sorting direction
          schema:
            example: DESC
            enum:
              - ASC
              - DESC
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              examples:
                Without pagination:
                  value:
                    - id: b1b73593-2445-421a-af42-359114d6c536
                      login: string
                      role: admin
                      createdAt: 1775287579998
                      updatedAt: 1775287579998
                  description: Without page and limit query parameters
                With pagination:
                  value:
                    total: 1
                    page: 1
                    limit: 10
                    data:
                      - id: b1b73593-2445-421a-af42-359114d6c536
                        login: string
                        role: admin
                        createdAt: 1775287579998
                        updatedAt: 1775287579998
                  description: With page and limit query parameters
        '400':
          description: Bad request. Wrong query parameters (hacker scope)
        '403':
          description: Insufficient permissions
      tags:
        - Users
  /user/{id}:
    get:
      operationId: UserController_findOne
      summary: Get single user by id
      description: Get single user by id
      parameters:
        - name: id
          required: true
          in: path
          description: User id, format UUID v4
          schema: {}
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              example:
                id: b1b73593-2445-421a-af42-359114d6c536
                login: string
                role: admin
                createdAt: 1775287579998
                updatedAt: 1775287579998
        '400':
          description: Bad request. User id is invalid (not uuid)
        '403':
          description: Insufficient permissions
        '404':
          description: User not found
      tags:
        - Users
    put:
      operationId: UserController_update
      summary: Update a user's password
      description: Updates a user's password by ID
      parameters:
        - name: id
          required: true
          in: path
          description: User id, format UUID v4
          schema: {}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePasswordDto'
      responses:
        '200':
          description: The user has been updated
          content:
            application/json:
              example:
                id: b1b73593-2445-421a-af42-359114d6c536
                login: string
                role: admin
                createdAt: 1775287579998
                updatedAt: 1775288465881
        '400':
          description: Bad request. User id is invalid (not uuid)
        '403':
          description: oldPassword is wrong or insufficient permissions
        '404':
          description: User not found
      tags:
        - Users
    delete:
      operationId: UserController_remove
      summary: Deletes user
      description: >-
        Deletes user by ID. Sets authorId to null on articles, deletes user's
        comments
      parameters:
        - name: id
          required: true
          in: path
          description: User id, format UUID v4
          schema: {}
      responses:
        '204':
          description: The user has been deleted
          example: ''
        '400':
          description: Bad request. User id is invalid (not uuid)
        '403':
          description: Insufficient permissions
        '404':
          description: User not found
      tags:
        - Users
...
info:
  title: Knowledge Hub
  description: Knowledge hub service for managing articles, categories, and comments
  version: '1.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
    CreateUserDto:
      type: object
      properties:
        login:
          type: string
          example: string
        password:
          type: string
          example: string
        role:
          type: string
          example: admin
      required:
        - login
        - password
    UpdatePasswordDto:
      type: object
      properties:
        oldPassword:
          type: string
          example: string
        newPassword:
          type: string
          example: string
      required:
        - oldPassword
        - newPassword
...